+++ packages/editor/src/utils/index.ts
@@ -3,7 +3,6 @@ export * from './indexMap';
 export * from './loader';
 export * from './nodeMap';
 export * from './color';
-export * from './color';
 export * from './common';
 export * from './contextMenu';
 export * from './create';
@@ -12,7 +11,6 @@ export * from './dom';
 export * from './eventEmiter';
 export * from './eventHelper';
 export * from './eventProxy';
-export * from './indexMap';
 export * from './init';
 export * from './inject';
 export * from './is';
